refactor: Lotto 일급 컬렉션 생성
